MDL-52172: Patch MNet installer to support submodules
Changes the upgrade/install part of MNet which sets the xmlrpcpath for a
published function to allow using a submodule style such as
mod/assign/submission/mahara, which currently gets stored as
assignsubmission/mahara which leads to errors when trying to call the
published methods.

This is fixed by changing the logic away from $modtype/$modname to just
the path of the module.

Change-Id: I9184c75c707085579514062b3db69be8f8b1d405